Failure handling for dsf updates
Summary:
On failure
- Teardown existing connection. So no more updates will
come on this subscription
- Cancel any outstanding updates (reset nextDsfUpdate pointer)
- Setup subscription again.

This is similar to previous behavior where we were performing
updates synchronously. There we would just teardown and resetup
the subscription to trigger a full sync. Now we need to clear out any
pending updates in the queue as well.

Reviewed By: peygar

Differential Revision:
D61254687

Privacy Context Container: L1125642

fbshipit-source-id: c4c80822f84ad297cee547eca1c59bbfec092747