Add gflag to bypass subscriber uniqueness
Summary:
I want to be able to bypass the dup subscriber error in tests just so they dont pollute our benchmarking.

This is just a temp solution to only be enabled in testing/debugging. I'd like to replace this with either
A. tcp level heartbeats, which seem to come to us more reliably than stream level heartbeats
B. a proper mechanism to bypass this using a client side flag. That requires us to untagle a bit of a web since we'll want to boot the old subscription out, which we currently have no mechanism to do so

Differential Revision:
D61315052

Privacy Context Container: L1125642

fbshipit-source-id: 90e90beaefc85219f98643afb3a6154484cb208a