Make sure to close publisher properly when GR is requested
Summary: I missed this when implementing patch publisher. When we send a request for gr, we queue an empty patch chunk and let this check break out of the publish loop. Right now we end up waiting for this for ever and crash on warmboot exit

Differential Revision:
D61443485

Privacy Context Container: L1125642

fbshipit-source-id: 2297df95d13a2784cc42be2f54a0fc67a6917869