move sai mirror manager to use port descriptor
Summary:
As titled, move sai mirror manager to use port descriptor to derive the monitor port.

For certain platforms, the monitor port will be a physical port
For DSF platforms, the monitor port can be a physical port (eventor port) or a system port. The system port can be a local or a remote system port.

Just derive the system port handle, derive the adapter key and program the monitor port.

Reviewed By: jasmeetbagga

Differential Revision: D61443902

fbshipit-source-id: 8f13b8c2d67b148a4ad2459076733b6ef89c940e