Back out "Default value for Port attr Reachability Group"
Summary:
Original commit changeset: 9f5bcb7f8305

Original Phabricator Diff: D61458742

This is due to the return code of the GET calls for reachability group list - we have an empty list of reachability group list (size 0). However, SDK is returning is buffer overflow, which should happen when size is larger than 0 and SDK requests fboss to allocate more space for the list attribute.

Will need Brcm to provide the fix to return 0 in this case, as the list size is 0.

Reviewed By: jasmeetbagga, srikrishnagopu

Differential Revision: D61545637

fbshipit-source-id: 44df60fc47241cbeb113f4fe5c98b31739b1dafc