Add port 65 to the BSP mapping
Summary: DVT units now have 65 ports (new management port was added). Port 65 was missing from the platform mapping, so we weren't tracking this port in qsfp_service. This was causing some issues in CLIs that expected this port to exist in the qsfp CLIs (since it exists in the agent CLIs)

Reviewed By: harshitgulati18

Differential Revision: D61502040

fbshipit-source-id: 289463fb12669b37e4f8fc3a5b642f6c2ef310a5