Enhance AclEntry switch config for multiple UDF match keys
Summary:
Newer requirements from backend T198076073 and https://fburl.com/workplace/wqekxutx require matching on 2 different offsets in a packet using UDF.
The current config allows for specifying multiple UDF groups to match in an ACL but lack the space to also specify the match data values.

Specifically,
```
  31: optional list<string> udfGroups;
  33: optional list<byte> roceBytes;
  34: optional list<byte> roceMask;
```
roceBytes and roceMask only allow to specify values for 1 udf group.

This change enhances this by grouping the group name and values into a single struct.
```
struct AclUdfEntry {
  1: string udfGroup;
  2: list<byte> roceBytes;
  3: list<byte> roceMask;
}
```
The existing config fields will be deprecated in favor of this new struct.

This also aligns with how SAI specifies values for UDF groups although more changes will be required to the ACL table configuration. That is out of scope for now.

Reviewed By: msomasundaran

Differential Revision: D61133956

fbshipit-source-id: e2eb80168b21efe6310d1ff1db86f6fdb40f16be