Change SubscriberChunk to always have a list of patches per key
Summary:
SubscriptionKey is effectively an alias (decided by the client) for each path that is being subscribed. In the raw path world this will always map to a single patch, but in the wildcard world a single path will resolve to multiple subscriptions and therefore multiple patches can come through.

So we need to have a list of patches per SubscriptionKey. Initially I was planning on having another chunk field dedicated for extended subs, but at this point that seems like more logic in Subscription.cpp than needed. Instead I'm just opting to always have a list, but for raw we know there will always be exactly 1.

Since no one is calling this api right now, its safe to swap out

Reviewed By: wilsonwinhi

Differential Revision: D61291174

fbshipit-source-id: 2c2782087df87b255346eac097efd55331959ace