Handle race in sys port, rif consumption exposed by async updates
Summary:
Cache current state of sysports and intfs received from remote
node. Since after the initial update we are not guaranteed to
receive the entire set of sysports and rifs in any update. For
instance on a nbr change, we will get the entire set of RIFs
but not sys ports. However when applying the update, we
apply RIFs and sys ports together. To that end we need to
cache the current sysports and rifs and then patch the
receieved update on top of it.
TODO: kill this code after we cutover to patch subscriptions

Reviewed By: shri-khare

Differential Revision:
D61645734

Privacy Context Container: L1125642

fbshipit-source-id: be174a24c55692a0905d9eab7443434152c39aba