Increase the thread heartbeat timeout
Summary: OSFP optics are taking a little longer to program. The delay gets long and over the current hearbeat threshold sometimes when we have 8 ports to program in an optic. This causes qsfp tests to fail. Increasing the timeout for now until we can optimize the programming on these optics (if at all it is possible)

Reviewed By: xanabani

Differential Revision: D61987557

fbshipit-source-id: 29495c83b93105b850a78cf0cde7cac067d72e86