Pick non loopback ports for verifyTxRxLatches
Summary: On meru800bia, mgmt port is looped back to itself. This causes verifyTxRxLatches to fail as it expects RX latches to report true when TX side is shut. In looped back port, both TX and RX sides are the same which fails the test

Reviewed By: xanabani

Differential Revision:
D62117477

Privacy Context Container: L1125642

fbshipit-source-id: ddca07711ec0abe352da3c488bc18a1ff58c9cdc