DLB: Hw test for zero length wr-imm packet
Summary:
Similar test as prior diff but this time for write immediate opcode for 0 length dma.
This one matches on
1. BTH OPCODE (17)
2. RETH dma length (0)
The dma length field is actually 4 bytes long. We will only match on LSB 2 bytes since packet length will never be longer than 4K for backend.
PCAP in this T198076073 for reference.

Differential Revision: D61730020

fbshipit-source-id: de330b6087a763ac18e26b629b9860774d57bad6