Enable Parsing ProductInfo for BMC-Lite platforms
Summary:
Fields from BMC-Lite platforms have to match the required fields from PlatformProductInfo.
The goal of this diff stack is to enable searching via Old BMC keys (e.g. "Facebook PCBA Part Number") and Keys from BMC-Lite EEPROM V5 (e.g. "Meta PCBA Part Number") in the fruid.json.

We will (eventually) use the output of weutil --json (or equivallent call in Platform Manager) as the source of information to initialize the system, and the keys in BMC-Lite are a little different

Reviewed By: mikechoifb

Differential Revision: D60961550

fbshipit-source-id: abf366695f1afa5105038d26423c7428b7d893f9