Add caching to PlatformNameLib
Summary:
Change getPlatformName to read from a specified file as a cache for its result, and getPlatformNameFromBios to write to this cache conditional on `writeToCache` parameter.

Only PM writes to the cache by making a call to `getPlatformNameFromBios` directly, to ensure that there are no racing writes.

The return value (ie platform name) is generally expected to be constant.

This reduces our direct dependency on `dmidecode` on the initialization of any of our config-driven services; `dmidecode` relies on kernel code which reads BIOS information and exposes it via sysfs.

Reviewed By: Scott8440

Differential Revision: D61576290

fbshipit-source-id: 35d61a8b17f334ca468343e6e01400cbd8c7ca13