Fix OSS runs of PlatformNameLibTest
Summary:
Example of failing test run: https://github.com/facebook/fboss/actions/runs/10802262654/job/29964924332

The root causes seems to be a difference in behavior between buck test and simply running the binary (i.e. buck run).

I'm not 100% sure what buck test is doing under the hood, but I'm guessing that it's running tests individually so that it can be distributed.

On the other hand for OSS runs and for simply executing the test locally via `buck run`, the `GetPlatformNameCache` test will fail because the stats are not reset, therefore the counter starts at 1 due to the previous increment during `GetPlatformNameNoCache`.

This change simply resets the stats at the start of the cache test.

Reviewed By: kimdo8736

Differential Revision: D62509385

fbshipit-source-id: 6910a26a2f24822e935fb793412bc1d669657b38