fan_service : Remove tests that does not make sense anymore.
Summary:
As in D62211686, presenceSysPath can be empty now, so the check in line 89 before this change can throw exception once we start using gpio, instead of sysPath.
Therefore we are removing this line that may cause problems - in D62211686, we are already checking whether fan has presenceSysPath or GPio, so this logic is already covered in the new diff (redundant. only causing problem. so removing)

Reviewed By: alandau

Differential Revision: D62535367

fbshipit-source-id: abe9819f8c4da8013511c025030e134ab285ccee