fboss2 show lldp
Summary:
As titled, today we derive expected LLDP peer from port description using
regex. And we have have added regex for few roles. And for those that we don't
have regex we print EMPTY.

Here is the solution that I am proposing and implementing, based on the
rationale that expected peer should reflect what is there in the config.
1. During show lldp, we anyways query agent to get PortInfo.
2. On receiving thrift call to populate PortInfo, let agent populate the
expected LLDP value(PeerName and PeerPort) values stored in Port as well.
3. Just display that.

Note: I am populating expectedLLDPPeerPort in peerInfo, but not displaying yet, will take it up later if needed!

Differential Revision: D62606002

fbshipit-source-id: fc96f486258625dbad82f8ceeb36f541810684ff