Add clang-format github workflow (#235)
Summary:
Pull Request resolved: https://github.com/facebook/fboss/pull/235

There have been quite a few imported PRs which end up triggering the clang format linter on sandcastle. The actual error is only visible to FB employees, while the original PR author can only see that a FB-internal CI job failed, and that only happens after the PR has already been imported.

This causes a lot of churn since we have to go back and forth.

Add a github workflow which replicates the sandcastle CLANGFORMAT linter. `.clang-format` file was copied from fbsource, which is the one that we use since we didn't have any fboss specific file which overrides.

Reviewed By: alandau

Differential Revision: D62764920

fbshipit-source-id: cc9b3b9c41cd6744e1f49039b4a18979ab9fbdc8