ip-address migration with missing rule for previous address.
Summary:
During ip address migration on SSW, we hit a condition where the rule was missing in the kernel table for the previous address. When we try to replace it with new address, we remove the previous rule and address and add the new one. As the previous rule was not existing it was causing wedge_agent crash.

Fix here is to ignore the rule removal if it does not exist.

Reviewed By: msomasundaran

Differential Revision:
D62905031

Privacy Context Container: L1125642

fbshipit-source-id: ef801507b494105c81a57b1244e57733a754f0d8