Changed type some params in thrift
Summary:
As titled. Since we do not send this data over the wire. And also this not something that is transferred between processes/services as IPC. I think this is safe change.

How we use these thrift fields.
1. In cfgr we read SI settings CSV provided by vendor and write those to
plaform mapping JSON.
2. We manually copy that generated plaform mapping JSON file to FB code in a
.cpp file.
3. When fboss agent process comes up, it loads the content in platform mapping.
JSON into this thrift.

And the reason for this change now is, new SI settings provided Cisco, has
values that doesn't fit in i16.

Reviewed By: harshitgulati18

Differential Revision: D62921959

fbshipit-source-id: c4d76a314b14d2f3759a0a26ffce0ac58ec738ae