Fix unused variable warning treated as an error for XGS OSS builds
Summary:
`matchV6` is an unused variable if performing an XGS build. This change adds guards to exclude the `matchV6` definition line in the case that we're performing such builds.

Similar story for `switchId` on line 1595 in ThriftHandler.cpp, except that it seems to be generally unused.

Reviewed By: srikrishnagopu

Differential Revision:
D62785477

Privacy Context Container: L1125642

fbshipit-source-id: 66dd6a5c5b78c378f4428b2bfe87f7d27250bb40