Use all ports on module when tcvr id is specified
Summary:
If we specify the transceiver ID, e.g.

```
sudo wedge_qsfp_util 20 --tx_disable
```
We currently will only operate on the first port on the module. If specifying the module ID, it seems like the intent would be to operate on the whole module, rather than just the first port on it.

This caused some issues in DNE testing: T202508655

Reviewed By: xanabani

Differential Revision: D63272105

fbshipit-source-id: dec878725e3d9a9e6dbcb8e69600d1488924f511