Refactor target translation own bzl file [4/7]
Summary:
There are some cases where the OSS target paths do not match the
internal target paths, and we need to translate the targets in OSS projects in
order to build successfully. This diff refactors all the logic related to this
translation into it's own file, so it is easier to reason about and doesn't get
mixed up with the shim implementations.

Reviewed By: stepancheg

Differential Revision: D63295974

fbshipit-source-id: 40274c190367566b562af9bca4e0dbaea2822b3e