Don't accidentally create intermediate nodes during remove
Summary:
Issue raised by NSDB where when SA publishes an empty delta for a device that does not exist in NSDB, we end up creating the state instead of a no-op.

This is because in the remove operation we will modify every intermediate node up to the path we want to remove. This is needed in case our state is published. But modify naively also creates the node if it does not exists, which in this case we don't want.

Fixing this by adding an option in modify to allow construction or not. In the case of `removePath` we never want construction

Differential Revision:
D63345154

Privacy Context Container: L1125642

fbshipit-source-id: 8c82f17c3d72cffdfd3c485aaa8668dd693d5e60