Start adding stats for load bearing port in errors, fec uncorrectables
Summary:
A link is load bearing as long as
- Its not drained
AND
- Its peer is not drained.
On such ports L1 errors are of deep interest and must be looked at.
However, with our current reporting its not easy to filter these out
on our dashboards. So add new counters which precisely capture these.
On DNX we will gather this info like below
- Local drained - port is drained or device is drained
- Remote drained - port is INACTIVE

If neither of these conditions are true, then in/fec errors are
impacting prod traffic and should be looked at. For non DNX platforms,
we will need to start exchanging this info via LLDP to gather this info.
L1 debug flow would be
- Figure out why such links were not drained by SH
- Figure out other more granular counters for (e.g. leaky bucket flaps, pre FEC BER
etch). We are not looking to fork all those counters separately

Reviewed By: srikrishnagopu

Differential Revision: D63658731

fbshipit-source-id: 630115466cecb0da707a7b2ebf00f185a1110878