rackmon2: Add a find API to RegisterMapDatabase
Summary:
This adds a find() API to the database class which
returns a const reference to the entires it contains.
This allows a user to request for all register maps
which match a given address instead of getting just
the one.

Keep the at() method and make it call the find()
interface.

Imported from: D63417448

Reviewed By: alandau

Differential Revision: D63711640

fbshipit-source-id: c1065ad9d5419145f9e3b97271703850d4615998