Avoid double frees from gpiod library
Summary:
gpiod library frees all lines associated with a chip when calling `gpiod_chip_close()` `GpiodLine` also frees the line when it is destroyed. Previously, we would close the chip before the GpiodLine went out of scope, which caused segfaults, but only on OSS builds (another problem which we will investigate further).

We plan on improving the usage of the gpiod library by using a GpiodChip wrapper later, but for now we solve the problem by ensuring GpiodLine goes out of scope before the chip is destroyed. This will unblock Celestica.

Reviewed By: alandau

Differential Revision: D63484887

fbshipit-source-id: 6b5a98d5fab418291a8ecbad26d2494b4402a8f8