HybridNode: add type tests for set, list, union, struct members
Summary:
In TestStruct add hybrid members with set, list, union and struct to start
building UTs for recurse visitor handling of hybrid nodes.

Note that in this diff allow_skip_thrift_cow annotation on these members is not
enabled. We will enable this after RecurseVisitor is updated to handle all
these types of hybrid members.

Reviewed By: wilsonwinhi

Differential Revision: D63838249

fbshipit-source-id: bb95ba200d830c5b833276000a76bd97412ee3d5