Fix firmware version publishing for case where hwmon exists but version files are not under it
Summary:
On Tahan, the version files for TAHAN_SMB_CPLD are not located under the `hwmon` directory, even though that directory is present. Firmware version publishing previously assumed that this would never happen. This change fixes that by selecting between `hwmon` and the base directory based on the presence of version files.

Since existence is now checked upfront, a new ODS key `ERROR_FILE_NOT_FOUND` is added to signal the case of no version files found in any of the checked directories.

This diff also removes ODS keys which encoded firmware versions as integer values, as we're not currently using them and it's adding some complexity.

Reviewed By: mikechoifb

Differential Revision: D63500511

fbshipit-source-id: 97c451807d99e1ff2b3950ead9f2c74cd7b21b4c