Add option to mark CONNECTED only on initial sync
Summary:
In Dsf we use SubscriptionState::CONNECTED to determine that we actually have all the state we need.

Previously, fsdb would serve heartbeats after serving initial sync so first chunk revieved would always be initial sync and we'd go into CONNECTED. But now heartbeats are served on a dedicated thread to avoid starvation, so we need to make sure the client doesnt prematurely mark CONNECTED.

Ideally we can always follow this behavior but actually switch agent relies on the reverse (CONNECTED to mean any chunk came through). For now introducing a flag to allow both behaviors, perhaps we can introduce a new state SYNCED to differentiate

Differential Revision: D63554533

fbshipit-source-id: d260d48b1d7dea4b02a7c5caed6fd77d11222456