gflags to enable firmware upgrade support on coldboot/link down
Summary: We are now changing the implementation to trigger firmware upgrade via a thrift call. Until that support is completely in place, guarding the existing support with gflag so that we don't lose it. This gflag will be disabled by default and enabled only in hw tests.

Reviewed By: birdsoup

Differential Revision:
D63620366

Privacy Context Container: L1125642

fbshipit-source-id: 3b8bf2170f1a88d8faa4256161e39f7e84586b2f