TransceiverManagerTest changes to pass in OSS binary run
Summary: This is a pre-req change to add QSFP service unit tests to github actions. Running the test binary and testing via `buck test` are slightly different because a raw run of the test binary will keep some state between tests, which doesn't appear to happen when using `buck test`, so we need to ensure that this flag is not forced to zero from the prior test (since that will prevent the service from warm booting).

Reviewed By: harshitgulati18

Differential Revision: D64063951

fbshipit-source-id: 47507bcfa49f7daa4b0e7e4bf7f7354cd8021cb2