CmdShowFabric: explicit column to identify mismatch
Summary:
Today we highlight this via color: Green on Match and Red on Mismatch.

But, this is not grep friendly. Colors are suppressed (by design) on grep.
The number of fabric ports is large, further making it harder to use visual
cue.

Thus, print an explicit column for match/mismatch that can be grep'ed.

Reviewed By: harshitgulati18

Differential Revision: D64131406

fbshipit-source-id: 6a27b404bb2508b74d186757c4465e6216eb6931