Fix setting maintap in zero preemphasis setting - second try
Summary:
Original diff D58880165 - we had issue on TH3 roundtrip tests, which got clarified during discussion with Brcm. Issue will be addressed in D64218169.

As titled. The goal is to set preemphasis + all six-tap values to zero (except main), and then set main to zero.

This is due to the check on yamp platform -
  ERROR: in BLACKHAWK, main TAP is expected to be greater than all other TAPs.ERROR:chip/blackhawk/tier1/blackhawk_tsc_config.c->blackhawk_tsc_validate_txfir_cfg() SerDes err_code = ERR_CODE_BAD_PTR_OR_INVALID_INPUT : 26
Fix this by first setting sixtap but keeping main the same, and then set main to zero.

Differential Revision: D61736113

fbshipit-source-id: fbcdbf1b300a7bf2c4c9e70f5cd7bc57416fd0ce