Modify expected TxFir setting if zero preemphasis is enabled - second try
Summary:
Original diff D57993788 - we had issue on TH3 roundtrip tests, which got clarified during discussion with Brcm. Issue will be addressed in D64218169.

As titled. Currently we set preemphasis to zero for port serdes, but did not maintain the zero preemphasis setting for next delta processing - this leads to resetting txFir to the desired values from platform mapping.
Fix this by
1. Process change zero preemphasis setting prior to serdes programming.
2. Different handling for leaba and brcm
a. Leaba requires remove and create for zero preemphasis
b. Brcm requires direct setting of zero preemphasis but not accepting create.
3. For desired TxFir setting, if zero preemphasis is set, all Tx setting should be zero.

Differential Revision: D61736115

fbshipit-source-id: 888857e37197ed0cbbb1432ca9078777b23f90f4