Consider switch drain state when computing the blink state of LEDs
Summary:
Today we only look at the portDrainState which will only be set when specific ports are drained. This doesn't indicate if the entire switch is drained. We need to account for switch drain state also so that we can blink LEDs indicating that the port is safe to operate or not.

Overall we blink the LED when

1. port is inactive (other side, either port or the entire switch, is drained)
2. local port state is drained (local port is drained)
3. local switch state is drained (local switch is drained)

Reviewed By: shri-khare

Differential Revision:
D64372511

Privacy Context Container: L1125642

fbshipit-source-id: 97707b5ba40a246af58673f05e9775b719077a20