Fix crash during subscription unregister
Summary: Calling cancelAndJoinAsync can cause a crash. Our thread has calls to a virtual method (serveSubscriptions), but by this point the subclass is destroyed so trying to join ends up resulting in a dump because of "pure virtual method called". Noticed this in DSF at high session scale when remote agent is restarting and unsubscribing

Differential Revision:
D64362881

Privacy Context Container: L1125642

fbshipit-source-id: f792ea82e9bce083873c08d6e8be0746228b47cd