Introduce ConnectionOptions
Summary:
We've collected a bit of of tech debt at low level thrift client. I'm trying to rework this to make it a bit easier to deal with. Mainly want to make it easier to pass new options down to this layer, since we need to make some of these timeouts tunable. Also want to make this stuff better unit testable

Starting with formalizing a ConnectionOptions class. This will include everything we need to get a socket up. This will replace ServerOptions, which is not really well named

Differential Revision: D64264977

fbshipit-source-id: 04719888fe4a6a8ff234ee334bbc892b52d336c7