(4/N): Remove restartSessionsAndExit method from fboss2 CLI
Summary:
As titled.

`restartSessionsAndExit` is the API BGP++ today opens to make it possible for one RPC call to send SIGTERM signal to BGP and restart the sessions.

This is not ideal to open the door for a unnoticed RPC call to nuke the BGP daemon. This is the subsequent change to remove all of the call sites for this API after deprecating the cold-boot API change. For the BGP restarting use case, the operational tooling today will invoke this via `systemctl` cmd

Reviewed By: shih-hao-tseng

Differential Revision: D64873876

fbshipit-source-id: 1f1c867b19b7b5c818099a30f293a63c4e360e2b