rename the snooping test argument in sendPkt function
Summary:
original argument is a on/off switch, if it's on,then snoop test enable. If expected packet received, then pass test. If it’s off, snooping test disable.

new argument indicate if packet is expected or not and snoop test always enable

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: daiwei1983

Differential Revision: D64765097

fbshipit-source-id: 52775768af4cb35986bfaf3618d92ed40eda63e2