Adding API to check isUpdateHw Protected in macTableHandler
Summary:
As per title, adding API to check if update hw protected

The API is called isUpdateHwProtected() and returns true if the platform supports hardware protection for MAC table updates. The API uses the transactionsSupported() API to check if the platform supports hardware transactions and returns true for SAI switches.

Reviewed By: msomasundaran

Differential Revision: D64576783

fbshipit-source-id: cfbe227dfe7a82a62d51256905ae5054ad877cee