adding resource check for MAC table in ResourceAccountant
Summary:
This diff adds a resource check for MAC table in the SwSwitch class. The code changes include adding a check for L2 resource in the SwSwitch class and modifying the ResourceAccountant class to include a check for L2 resource. The ResourceAccountant class also includes a getL2Entries() function to get the number of L2 entries programmed.

The changes are made to ensure that the MAC table does not exceed the maximum number of entries supported, which is 1000 .

Reviewed By: zechengh09

Differential Revision: D63394006

fbshipit-source-id: 476a4855934f6baf03372350f62fb9f00f245a9e