Enable packaging of binaries built from local changes
Summary: Allow packaging of FBOSS binaries built from a local commit. In these cases, the git path does not exist because the local checkout of the source code is used rather than re-cloning via getdeps, so the path from which to copy the configs/run scripts needs to be adjusted.

Differential Revision: D65085740