Print Link Test Production features
Summary:
Added the `printProductionFeatures` and `getProductionFeatures` functions to the `LinkTest` class. The `printProductionFeatures` function outputs the `production features` of a specific test (L1/L2) when the `list_production_features` flag is enabled. The `getProductionFeatures` function is a virtual method overridden by each test class that inherits from `LinkTest`. Since some gtests directly use LinkTest, it also implements the `getProductionFeatures` function instead of leaving it as a pure virtual method.

Here's a detailed list of the current Link tests, including their type, file name, and Test class, to facilitate cross-checking the changes in this diff.

https://docs.google.com/spreadsheets/d/1MorIYHYDufAmjEC6om4V5T1KPXGIZwudiYqDlTyql9Q/edit?usp=sharing

**Note that:**
 1) Most Test classes are straightforward, with all tests being homogenous (e.g., MacsecTest). In these cases, we simply override the getProductionFeatures function in the respective class.

2) However, there are mixed cases like LinkTest and LinkSanityTestDataPlaneFlood. In these instances, we determine the type of Production feature to return based on the test name.

Reviewed By: xanabani

Differential Revision: D64547239

fbshipit-source-id: 212164e6b09baaad257cf03f1e8f0747067ad6a3