Enable packaging of binaries built from local changes (#283)
Summary:
Pull Request resolved: https://github.com/facebook/fboss/pull/283

Allow packaging of FBOSS binaries built from a local commit. In these cases, the git path does not exist because the local checkout of the source code is used rather than re-cloning via getdeps, so the path from which to copy the configs/run scripts needs to be adjusted.

Reviewed By: srikrishnagopu

Differential Revision: D65085740

fbshipit-source-id: d4c309662805109657acf1c8a5d24cdc472a3e1e