Fix PG lossless mode programming
Summary:
This bug is discovered as part of the investigation of S455066.
```
    OPERATIONAL_STATE=VALID
    PFC=1
    LOSSLESS=1
    ING_MIN_MODE=USE_PRI_GRP_MIN
    TM_PRI_GRP_ID=6
    PORT_ID=0x11(17)
```
Above is an example from a production RTSW where PG6, which is lossy with headroom limit 0 but has LOSSLESS=1.

A PG can be lossless only if it has non zero headroom limit. The existing code just checks for the presence of headroomLimit since it is an optional field but not the value. For lossy PGs, headroom limit is 0 (most if not all cases) in which case it cannot be programmed as lossless.

Just fixing the code is not a solution since changing the lossless mode with live traffic is not supported by SDK and can corrupt the MMU. FLAGS_fix_lossless_mode_per_pg will be used in conjunction with a CFGR feature flag to make it a disruptive config change.

Reviewed By: nivinl

Differential Revision: D64864808

fbshipit-source-id: c56d4d4740239833b9250900f27ffa0ee407cb40