Raise Runtime Exception for Link Tests with No Production Features
Summary:
We raise a ***runtime exception*** when a link test does not produce any production features, the goal is to ***enforce each Link test to have a production feature***.

Note that, raising the exception will not immediately stop the execution of gtests. The individual test will be marked as failed and other tests will be still marked as skipped. ***But the test suite will return a non-0 exit code***.

Netcastle expects a 0 exit code when running tests with the ***--list_production_feature flag***. If the exit code is non-0, Netcastle will exit with an exception. So it fits the bill nicely.

Similar change and reasoning for agent ensemble link tests.

Reviewed By: xanabani

Differential Revision: D65465279

fbshipit-source-id: 7c3dbf1db2e5fd00e376bcfb801cf019d381143f