Minor fix in ACL table recreate sequence
Summary:
During canary-on/off, if table recreation is required, the existing logic
- deletes ACL entries in table
- delete table
- add table
- Add all ACL entries back

A problem with step4 above is there may be ACL entries which are removed in configuration and also incompatible with the new table. An example is an  ACL entry maybe reliant on a qualifier which is removed during table creation.

The fix is to only add ACL entries which are present in the new ACL table. All new ACL entries will be added in later ACL entry delta processing as before.

Differential Revision: D65558887

fbshipit-source-id: 426bb67cbe597bcd9cf50d073680722495ccca34