HW asic config golden test
Summary:
Add a test to verify default register/memory programming for J3AI.

We first run the test with `--dump_golden_data` to dump default ASIC programming and check it in as `fboss/agent/hw/test/golden/asic/<asic>.csv`. If a change is detected due to SDK changes etc, the differences should be manually reviewed for acceptance / rejection.

Reviewed By: nivinl

Differential Revision: D63490679

fbshipit-source-id: 7531e6b725877e0a0129181e50e22e6bc39c8c0e