Fix show lldp to show the proper peer
Summary:
if show lldp cli uses extractExpectedPort() API, for fsw peers, it shows up as below

```
devvm7013:~$ fboss2 -H rsw002.p001.m001.qzs1 show lldp
 Local Int  Status  Expected-Peer                           LLDP-Peer              Peer-Int   Platform  Peer Description
------------------------------------------------------------------------------------------------------------------------------------------------
 eth1/1/1   up      fsw001.p001.m001.qzs1:eth2/13/1.fsw001  fsw001.p001.m001.qzs1  eth2/13/1  FBOSS     fsw001.p001.m001.qzs1:eth2/13/1
 eth1/5/1   up      fsw001.p001.m001.qzs1:eth2/14/1.fsw001  fsw001.p001.m001.qzs1  eth2/14/1  FBOSS     fsw001.p001.m001.qzs1:eth2/14/1
 eth1/9/1   up      fsw002.p001.m001.qzs1:eth2/13/1.fsw002  fsw002.p001.m001.qzs1  eth2/13/1  FBOSS     fsw002.p001.m001.qzs1:eth2/13/1
 eth1/13/1  up      fsw002.p001.m001.qzs1:eth2/14/1.fsw002  fsw002.p001.m001.qzs1  eth2/14/1  FBOSS     fsw002.p001.m001.qzs1:eth2/14/1

devvm7013:~$
```

this is because the first split on " " populates result[0] with the entire string and a subsequen split on "." populated result[1] with "fsw001". so this ends up adding "eth" portion of the string and an unncessary "fsw001" in the end and cli showing a mismatch in peer from expectation.

Change the index to instead print the output as "fsw001.p001" so the expected and actual peer matches.

Differential Revision: D65976502

fbshipit-source-id: eceab1643820672407a6e906a941645789e67ce8