Helper function to update FEC tail in Phy Info
Summary:
FEC can correct anywhere from 1-15 codewords. FEC tail is the maximum codewords that it corrected.

We get the number of codewords count for each of the buckets from the hardware. While this is a useful number, fecTail is an easier metric to track and also a single value that we can publish to ODS rather than all the 16 codeword counters per port.

Thus in this diff stack, I add fecTail to PhyInfo, fb303 and eventually ODS.

Reviewed By: xanabani

Differential Revision:
D65649095

Privacy Context Container: L1125642

fbshipit-source-id: 663a675bbece9b2d6adcc2f07ef2450e452e32bc