Remove unused-variable in fboss/agent/test/fsdb_tests/facebook/FsdbSyncerTest.cpp +3
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: jasmeetbagga

Differential Revision: D66330542

fbshipit-source-id: 749857b92d9f33a0727019de912b6590e0381d62