Populate fec codeword stats and fec tail in BcmPort
Summary:
We still have native BcmPort applications in the fleet, hence updating fec codewords and fec tails in PhyInfo/fb303.

We should probably just get rid of fdrStats_ map and directly use the new codewordStats_ map. fdrStats_ stores a time series which makes it hard to get the overall accumulated counter for publishing to PhyInfo. I'll clean this up in a subsequent diff.

Reviewed By: jasmeetbagga

Differential Revision: D65674877

fbshipit-source-id: a95c79a2307845ee25e86b1fb8ed4796c9f51734