Remove unused-variable in fboss/agent/hw/HwBasePortFb303Stats.cpp +3
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D66330319

fbshipit-source-id: c5825139d116c8a2774e02221b7a44e6536687f2