Remove unused-variable in fboss/lib/phy/facebook/bcm/util/fboss-xphy.cpp +3
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D66330447

fbshipit-source-id: d139d857c185dd9630bacefe82d3a9240ac25f6b