add EnableHybridStorage compilation flag in ConvertToNodeTraits
Summary:
the original issue is that hybrid node conversion only works for the first layer. The further layer relies on `ConvertToNodeTraits` to determine the struct_type, which is the node type having the flag off as default. This further passes down the flag to deeper nodees.
note: Since `ConvertToNodeTraits` is using variadic template, i cannot pass EnableHybridStorage flag in the end, i need to pass that in the beginning to get around compilation issue.

Differential Revision:
D66402832

Privacy Context Container: L1125642

fbshipit-source-id: d58252ac4c33ca6550b0244a7b44b03860ac688e