store the first matched pciDev's sysfsPath in PciExplorer
Summary:
Although it's valid to have multiple PCIe devices with the same device IDs
in a single system, each FPGA/CPLD has its unique device ID in the FBOSS
environment (defined in the FBOSS System Requirements).

As a result, it's safe (and faster) to exit the PCIe sysfs enemuration loop
as soon as the first matched PCIe device is found.

Reviewed By: kimdo8736

Differential Revision: D66520727

fbshipit-source-id: 864d62d52f7c35160445d7314c08de8a8b15da8a