Rename getNumVoqs to getLocalPortNumVoqs
Summary:
As titled. `getNumVoqs()` should only be used for local ports, as it relies on the dual stage flag for local RDSW/EDSW.

Rename it to `getLocalPortNumVoqs` - we will add getRemotePortNumVoqs that's catering remote sys ports. This will be used when creating RCY ports for remote RDSW and EDSW.

Reviewed By: jasmeetbagga

Differential Revision:
D66680057

Privacy Context Container: L1125642

fbshipit-source-id: 2c2a2301435872b367b8b138658392286c5018d1