update SubscriptionStateChangeCb to indicate initialSync
Summary:
FsdbSubscriber may subscribe to a path that may or may not be (yet) published by Publisher. In such a case, when the subscription stream is connected, subscriber needs a mechanism to determine whether the dataCb will be invoked or not.

For above functionality, SubscriptionStateChangeCb is updated to include an optional<bool> indicating whether to expect the dataCb for initialSync. This will be provided only when the stream is connected, and only for Patch subscriptions.

Reviewed By: wilsonwinhi

Differential Revision: D66247020

fbshipit-source-id: 452cdb63bf3dfe8af7779900e081424e5f0a7fb6