Thrift support for local/remote VOQ max latency config
Summary:
Expose the max expected latency for local / remote L1 / remote L2 voqs, which will be
used to report latency watermark per VoQ. In case latency seen on VoQ is less than this
configured limit, then the configured limit will be reported as latency watermark. However,
if the latency seen on VoQ is more than this value, the value configured in "out of bounds"
gets reported.

In SAI< we have an option to configure separate "out of bounds" for local, remote L1 and
L2, however, don't see a need for us to expose all 3 with a config, any latency exceeded
case can report the same "out of bounds" value which should be good for our use case.

Reviewed By: jasmeetbagga

Differential Revision: D66047986

fbshipit-source-id: 6aacc93c70d2f5879425af067e4c8c97b2d84964