Fix VOQ latency watermark counter to be read on clear
Summary:
There is a bug in the way VoQ latency watermarks are read, these are not read as COR
counters as of now, 'coz of which, once the high value is seen, it always remains. Fixing
that by moving this counter to be read along with VoQ byte counter which gives the queue
watermarks.

Reviewed By: jasmeetbagga

Differential Revision: D66917087

fbshipit-source-id: a545b1aa473b331f137ae437d7e6849942fee6fd